azure: fixes instance id when kernel gets upgraded from a version lower than 4.17 to a version higher than 4.17
This is a dumb fix for only one situation (when the upgrade kernel
path is from a version lower than 4.17 to a version higher than
4.17.

Partially fixes: https://bugs.launchpad.net/cloud-init/+bug/1803173

It is a PoC to prove that cloud-init can fix the issue.

The real fix would have to handle downgrades too, meaning that the
instance-id check should be case insensitive at least on Azure and
it should be done at the framework level, not here.

For a downgrade, it should be .lower() instead of .upper()

The whole problem has been introduced by
https://github.com/torvalds/linux/commit/712ff25450bd01366301eef81c33e865d901e7b7#diff-f2bd14bc67b5e2da67116bca971bbd0b
where the dmi uuid was changed from uppercase to lowercase.

This was initially fixed in walinuxagent, but at some moment,
walinuxagent was "fixed" not to touch cloud-init data,
and the problem reapeared.